Add: Exception handler

+++ src/main/java/app/bys/bys_api/service/FinalUserService.java
@@ -6,6 +6,7 @@ import app.bys.bys_api.model.entity.FinalUser;
 import app.bys.bys_api.mapper.FinalUserMapper;
 import app.bys.bys_api.mapper.PageMapper;
 import app.bys.bys_api.repository.FinalUserRepository;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -18,7 +19,8 @@ public class FinalUserService {
     private final FinalUserMapper mapper;
 
     public FinalUserDto get(Long id) {
-        return mapper.entityToDto(finalUserRepository.findById(id).orElseThrow());
+        return mapper.entityToDto(finalUserRepository.findById(id).
+                orElseThrow(() -> new EntityNotFoundException("Final user with id: " + id + " not found")));
     }
 
     public PageDto<FinalUserDto> getAll(Pageable pageable) {
@@ -30,12 +32,16 @@ public class FinalUserService {
     }
 
     public FinalUserDto update(Long id, FinalUserDto finalUserDto) {
-        FinalUser userFound = finalUserRepository.findById(id).orElseThrow();
+        FinalUser userFound = finalUserRepository.findById(id).
+                orElseThrow(()-> new EntityNotFoundException("Final user with id: " + id + " not found"));
         mapper.updateFinalUserFromDto(finalUserDto, userFound);
         return mapper.entityToDto(finalUserRepository.save(userFound));
     }
 
     public void delete(Long id) {
+        if (!finalUserRepository.existsById(id)) {
+            throw new EntityNotFoundException("Final user with id: " + id + " not found");
+        }
         finalUserRepository.deleteById(id);
     }
 }
