+++ src/main/java/app/bys/bys_api/service/SpecializationService.java
@@ -6,6 +6,7 @@ import app.bys.bys_api.model.dto.PageDto;
 import app.bys.bys_api.model.dto.SpecializationDto;
 import app.bys.bys_api.model.entity.Specialization;
 import app.bys.bys_api.repository.SpecializationRepository;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -18,7 +19,8 @@ public class SpecializationService {
     private final SpecializationMapper mapper;
 
     public SpecializationDto get(Long id) {
-        return mapper.entityToDto(specializationRepository.findById(id).orElseThrow());
+        return mapper.entityToDto(specializationRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Specialization with id " + id + " not found")));
     }
 
     public PageDto<SpecializationDto> getAll(Pageable pageable) {
@@ -30,12 +32,17 @@ public class SpecializationService {
     }
 
     public SpecializationDto update(Long id, SpecializationDto specializationDto) {
-        Specialization specializationFound = specializationRepository.findById(id).orElseThrow();
+        Specialization specializationFound = specializationRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Specialization with id " + id + " not found"));
+
         mapper.updateSpecializationFromDto(specializationDto, specializationFound);
         return mapper.entityToDto(specializationRepository.save(specializationFound));
     }
 
     public void delete(Long id) {
+        if (!specializationRepository.existsById(id)) {
+            throw new EntityNotFoundException("Specialization with id " + id + " not found");
+        }
         specializationRepository.deleteById(id);
     }
